--- Flask_SQL/Server/Conv1D/RunSmallGA_Conv1D_Test.sh
@@ -1,16 +1,0 @@
-python3 RunGA_Conv1D.py \
-    --num_generations 50 \
-    --num_epochs 50 \
-    --mutate_prob 0.05 \
-    --population_size 100 \
-    --train_file exoplanet \
-    --max_vae_hidden_layers 2 \
-    --max_vae_latent 8 \
-    --max_dnn_latent 8 \
-    --max_vae_hidden 64 \
-    --max_dnn_hidden 64
-    --max_dnn_hidden_layers 2 \
-    --max_conv_layers 2 \
-    --max_kernel_size 2 \
-    --max_pool_size 0 \
-    --max_filter_size 64
