+++ Flask_SQL/Server/templates/index.html
@@ -124,8 +124,8 @@
                 <form style="z-index:1">
                     Node sorting:
                     <label><input type="radio" name="sort_type" value="fitness" checked="true">Fitness</label>
-                    <label><input type="radio" name="sort_type" value="scaled_size" >Size</label>
-                    <label><input type="radio" name="sort_type" value="scaled_height" >Height</label>
+                    <label><input type="radio" name="sort_type" value="scaled_size" >Num Conv</label>
+                    <label><input type="radio" name="sort_type" value="scaled_height" >Num DNN</label>
                     <label><input type="radio" name="sort_type" value="unsorted" >Index</label>
                 </form>
 
@@ -133,8 +133,8 @@
                 <form style="z-index:1">
                     Node colors:
                     <label><input type="radio" name="color_type" value="fitness" checked="true">Fitness</label>
-                    <label><input type="radio" name="color_type" value="size" >Size</label>
-                    <label><input type="radio" name="color_type" value="height" >Height</label>
+                    <label><input type="radio" name="color_type" value="size" >Num Conv</label>
+                    <label><input type="radio" name="color_type" value="height" >Num DNN</label>
                 </form>
 
                 <div id="evolution"></div>
@@ -165,8 +165,8 @@
         var format_number = d3.format( "4.2f" ) ,
             format = function( d ) {
                 str  = "Fitness: " + format_number( d.fitness );
-                str += " ,size: " + d.size + "(" + format_number( d.scaled_size ) + ")";
-                str += " ,height: " + d.height + "(" + format_number( d.scaled_height ) + ")";
+                str += " ,#Conv: " + d.size + "(" + format_number( d.scaled_size ) + ")";
+                str += " ,#DNN: " + d.height + "(" + format_number( d.scaled_height ) + ")";
                 return  str;
             } ,
             color = d3.scale.category20();
